Factor out embedding methods into module.
Move functionality only related to embeddings to embeddings module.
